API

The AutomaticHistogram type

AutoHist.AutomaticHistogram — Type

AutomaticHistogram

A type for representing a histogram where the histogram partition has been chosen automatically based on the sample. Can be fitted to data using the fit, histogram_irregular or histogram_regular methods.

Fields

breaks: AbstractVector consisting of the cut points in the chosen partition.

density: Estimated density in each bin.

counts: The bin counts for the partition corresponding to breaks.

type: Symbol indicating whether the histogram was fit using an irregular procedure (type==:irregular) or a regular one (type==:regular).

closed: Symbol indicating whether the drawn intervals should be right-inclusive or not. Possible values are :right (default) and :left.

a: Value of the Dirichlet concentration parameter corresponding to the chosen partition. Only of relevance if a Bayesian method was used to fit the histogram, and is otherwise set to NaN.

Examples

julia> using AutoHist

julia> x = LinRange(eps(), 1.0-eps(), 5000) .^(1.0/4.0);

julia> h = fit(AutomaticHistogram, x)
AutomaticHistogram
breaks: [0.00012207031249977798, 0.17763663029325183, 0.29718725232110504, 0.4022468898607337, 0.4928155429121377, 0.5797614498414855, 0.6667073567708333, 0.7572760098222373, 0.8405991706295289, 0.9202995853147645, 1.0000000000000002]
density: [0.00662683597412854, 0.057821970706400425, 0.17596277991076312, 0.36279353706969375, 0.6214544825215076, 0.9730458529384184, 1.4481767793920146, 2.0440057561776532, 2.733509595364622, 3.545742066060367]
counts: [5, 34, 92, 164, 270, 423, 656, 852, 1090, 1414]
type: irregular
closed: right
a: 5.0

Fitting an automatic histogram to data

An automatic histogram based on regular or irregular partitions can be fitted to the data by calling the fit method.

StatsAPI.fit — Method

fit(AutomaticHistogram, x::AbstractVector{x<:Real}; rule=:default, type=:irregular, kwargs...)

Fit a histogram to a one-dimensional vector x with an automatic and data-based selection of the histogram partition.

Arguments

x: 1D vector of data for which a histogram is to be constructed.

Keyword arguments

rule: The criterion used to determine the optimal number of bins. Default value is rule=:default which uses the default rule for the regular or irregular histogram procedure depending on the value of type.

type: Symbol indicating whether the fitted method is a regular and irregular one. The rules :bayes, :l2cv, :klcv and :nml are implemented for both regular and irregular histograms, and this keyword specifies whether the regular or irregular version should be used. For other rules, this function automatically infers the type from the rule keyword, and misspecifying the rule in this case has no effect. Possible values are :irregular (default) and :regular.

kwargs: Additional keyword arguments passed to histogram_regular or histogram_irregular depending on the specified or inferred type.

Returns

h: An object of type AutomaticHistogram, corresponding to the fitted histogram.

Examples

julia> x = randn(10^3)
julia> h1 = fit(AutomaticHistogram, x)                                      # fits an irregular histogram
julia> h2 = fit(AutomaticHistogram, x; rule=:wand, scalest=:stdev, level=4) # fits a regular histogram

Alternatively, an automatic histogram can be fitted to the data by the following methods:

AutoHist.histogram_irregular — Function

histogram_irregular(x::AbstractVector{<:Real}; rule::Symbol=:bayes, grid::Symbol=:regular, closed::Symbol=:right, alg::AbstractAlgorithm=:DP, maxbins::Int=-1, support::Tuple{Real,Real}=(-Inf,Inf), use_min_length::Bool=false, logprior::Function=k->0.0, a::Real=1.0)

Create an irregular histogram based on optimization of a criterion based on Bayesian probability, penalized likelihood or LOOCV. Returns an AutomaticHistogram object with the optimal partition corresponding to the supplied rule.

Arguments

x: 1D vector of data for which a histogram is to be constructed.

Keyword arguments

rule: The criterion used to determine the optimal number of bins. Defaults to the Bayesian method of Simensen et al. (2025).

grid: Symbol indicating how the finest possible mesh should be constructed. Options are :data, which uses each unique data point as a grid point, :regular (default) which constructs a fine regular grid, and :quantile which constructs the grid based on the sample quantiles.

closed: Symbol indicating whether the drawn intervals should be right-inclusive or not. Possible values are :right (default) and :left.

alg: Algorithm used to fit the model. Possible options are DP() and GDPD(). See DP and GPDP for further details.

maxbins: The maximal number of bins to be considered by the optimization criterion. Only used if grid is set to :regular or :quantile. Defaults to maxbins=min(4*n/log(n)^2, 1000). If the specified argument is not a positive integer, the default value is used.

support: Tuple specifying the the support of the histogram estimate. If the first element is -Inf, then minimum(x) is taken as the leftmost cutpoint. Likewise, if the second elemen is Inf, then the rightmost cutpoint is maximum(x). Default value is (-Inf, Inf), which estimates the support of the data.

use_min_length: Boolean indicating whether or not to impose a restriction on the minimum bin length of the histogram. If set to true, the smallest allowed bin length is set to (maximum(x)-minimum(x))/n*log(n)^(1.5).

logprior: Unnormalized logprior distribution for the number k of bins. Defaults to a uniform prior. Only used when rule keyword is set to :bayes.

a: Dirichlet concentration parameter in the Bayesian irregular histogram model. Only used when rule is set to :bayes.

Returns

h: The fitted histogram as an AutomaticHistogram object.

Examples

julia> x = [0.037, 0.208, 0.189, 0.656, 0.45, 0.846, 0.986, 0.751, 0.249, 0.447]
julia> h1 = histogram_irregular(x)
julia> h2 = histogram_irregular(x; grid=:quantile, support=(0.0, 1.0), logprior=k->-log(k), a=sqrt(10))

AutoHist.histogram_regular — Function

histogram_regular(x::AbstractVector{<:Real}; rule::Symbol=:bayes, closed::Symbol=:right, maxbins::Int=1000, support::Tuple{Real,Real}=(-Inf,Inf), logprior::Function=k->0.0, a::Union{Real,Function}=1.0, level::Int=2, scalest::Symbol=:minim)

Create a regular histogram based on an asymptotic risk estimate, or optimization criteria from Bayesian probability, penalized likelihood or LOOCV. Returns a AutomaticHistogram object with regular bins, with the optimal bin number corresponding to the supplied criterion.

Arguments

x: 1D vector of data for which a histogram is to be constructed.

Keyword arguments

rule: The criterion used to determine the optimal number of bins. Defaults to the method Bayesian method of Simensen et al. (2025)

closed: Symbol indicating whether the drawn intervals should be right-inclusive or not. Possible values are :right (default) and :left.

maxbins: The maximal number of bins to be considered by the optimization criterion. Ignored if the specified argument is not a positive integer. Defaults to maxbins=1000

support: Tuple specifying the the support of the histogram estimate. If the first element is -Inf, then minimum(x) is taken as the leftmost cutpoint. Likewise, if the second element is Inf, then the rightmost cutpoint is maximum(x). Default value is (-Inf, Inf), which estimates the support of the data.

logprior: Unnormalized logprior distribution of the number k of bins. Only used in the case where the supplied rule is :bayes. Defaults to a uniform prior.

a: Specifies Dirichlet concentration parameter in the Bayesian histogram model. Can either be a fixed positive number or a function computing aₖ for different values of k. Defaults to 1.0 if not supplied. Uses default if suppled value is negative.

scalest: Estimate of scale parameter used in computing Wands' rule. Only used if rule is set to :wand. Possible values are :minim :stdved and :iqr. Default value is scalest=:minim.

level: Specifies the level used for the Kernel functional estimate in Wands' rule. Only used if rule==:wand. Possible values are 0,1,2,3,4 and 5. Default value is level=2.

Returns

h: The fitted histogram as an AutomaticHistogram object.

Examples

julia> x = [0.037, 0.208, 0.189, 0.656, 0.45, 0.846, 0.986, 0.751, 0.249, 0.447]
julia> h1 = histogram_regular(x)
julia> h2 = histogram_regular(x; logprior=k->-log(k), a=k->0.5*k)

Additional methods for AutomaticHist

AutoHist.peaks — Method

peaks(h::AutomaticHistogram)

Return the location of the modes/peaks of h as a Vector, sorted in increasing order.

Formally, the modes/peaks of the histogram h are defined as the midpoints of an interval $\mathcal{J}$, where the density of h is constant on $\mathcal{J}$, and the density of h is strictly smaller than this value in the histogram bins adjacent to $\mathcal{J}$. Note that according this definition, $\mathcal{J}$ is in general a nonempty union of intervals in the histogram partition.

Base.minimum — Method

minimum(h::AutomaticHistogram)

Return the minimum of the support of h.

Base.maximum — Method

maximum(h::AutomaticHistogram)

Return the maximum of the support of h.

Base.extrema — Method

extrema(h::AutomaticHistogram)

Return the minimum and the maximum of the support of h as a 2-tuple.

Distributions.insupport — Method

insupport(h::AutomaticHistogram, x::Real)

Return true if x is in the support of h, and false otherwise.

Distributions.pdf — Method

pdf(h::AutomaticHistogram, x::Real)

Evaluate the probability density function of h at x.

Base.length — Method

length(h::AutomaticHistogram)

Returns the number of bins of h.

StatsAPI.loglikelihood — Method

loglikelihood(h::AutomaticHistogram)

Compute the log-likelihood (up to proportionality) of an h.

The value of the log-likelihood is ∑ⱼ Nⱼ log (dⱼ), where Nⱼ, dⱼ are the bin counts and estimated densities for bin j.

AutoHist.logmarginallikelihood — Function

logmarginallikelihood(h::AutomaticHistogram, a::Real)
logmarginallikelihood(h::AutomaticHistogram)

Compute the log-marginal likelihood (up to proportionality) of h when the value of the Dirichlet concentration parameter equals a. This can be automatically inferred if the histogram was fitted with rule=:bayes, and does not have to be explicitly passed as an argument in this case.

Assumes that the Dirichlet prior is centered on the uniform distribution, so that aⱼ = a/k for a scalar a>0 and all j. The value of the log-marginal likelihood is ∑ⱼ {log Γ(Nⱼ+aⱼ) - log Γ(aⱼ) - Nⱼlog(dⱼ)} - log Γ(a+n) + log Γ(a), where where Nⱼ is the bin count for bin j.

Base.convert — Method

convert(Histogram, h::AutomaticHistogram)

Convert an h to a StatsBase.Histogram, normalized to be a probability density.

AutoHist.distance — Function

distance(h1::AutomaticHistogram, h2::AutomaticHistogram, dist::Symbol=:iae; p::Real=1.0)

Compute a statistical distance between two histogram probability densities.

Arguments

h1, h2: The two histograms for which the distance should be computed

dist: The name of the distance to compute. Valid options are :iae (default), :ise, :hellinger, :sup, :lp. For the $l_p$-metric, a given power p can be specified as a keyword argument.

Keyword arguments

p: Power of the $l_p$-metric, which should be a number in the interval $[1, \infty]$. Ignored if dist != :lp. Defaults to p=1.0.

Index

AutoHist.AutomaticHistogram

AutoHist.distance

AutoHist.histogram_irregular

AutoHist.histogram_regular

AutoHist.logmarginallikelihood

AutoHist.peaks

Base.convert

Base.extrema

Base.length

Base.maximum

Base.minimum

Distributions.insupport

Distributions.pdf

StatsAPI.fit

StatsAPI.loglikelihood